Resove review by @wschella
