+++ lib/functions/Helpers.ts
@@ -9,8 +9,8 @@ import type { Literal } from '../expressions';
 import * as E from '../expressions';
 import * as C from '../util/Consts';
 import { TypeURL } from '../util/Consts';
+import type { IDateTimeRepresentation } from '../util/DateTimeHelpers';
 import * as Err from '../util/Errors';
-import type { IDateTimeRepresentation } from '../util/InternalRepresentations';
 import type { ArgumentType } from './Core';
 import type { ImplementationFunction } from './OverloadTree';
 import { OverloadTree } from './OverloadTree';
