+++ test/integration/functions/op.addition.test.ts
@@ -1,5 +1,5 @@
 import { TypeURL } from '../../../lib/util/Consts';
-import { dateTimeNotation, dayTimeDurationNotation, int, numeric } from '../../util/Aliases';
+import { dateTimeTyped, dayTimeDurationTyped, int, numeric } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import type { ITestTableConfigBase } from '../../util/utils';
 import { runTestTable } from '../../util/utils';
@@ -47,7 +47,7 @@ describe('evaluation of \'+\' like', () => {
       "0"^^xsd:double 0d = "0.0E0"^^xsd:double
       "0"^^xsd:double 0f = "0.0E0"^^xsd:double
       
-      '${dateTimeNotation('2012-02-28T12:14:45Z')}' '${dayTimeDurationNotation('P2D')}' = '${dateTimeNotation('2012-03-01T12:14:45Z')}'
+      '${dateTimeTyped('2012-02-28T12:14:45Z')}' '${dayTimeDurationTyped('P2D')}' = '${dateTimeTyped('2012-03-01T12:14:45Z')}'
     `,
   });
   runTestTable({
