+++ test/integration/functions/op.day.test.ts
@@ -1,4 +1,4 @@
-import { dateNotation, int } from '../../util/Aliases';
+import { dateTyped, int } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'DAY\'', () => {
     notation: Notation.Function,
     operation: 'DAY',
     testTable: `
-    '${dateNotation('2010-06-21Z')}' = '${int('21')}'
-    '${dateNotation('2010-12-21-08:00')}' = '${int('21')}'
-    '${dateNotation('2008-06-20Z')}' = '${int('20')}'
-    '${dateNotation('2011-02-01')}' = '${int('1')}'
+    '${dateTyped('2010-06-21Z')}' = '${int('21')}'
+    '${dateTyped('2010-12-21-08:00')}' = '${int('21')}'
+    '${dateTyped('2008-06-20Z')}' = '${int('20')}'
+    '${dateTyped('2011-02-01')}' = '${int('1')}'
   `,
   });
 });
