+++ test/integration/functions/op.hours.test.ts
@@ -1,4 +1,4 @@
-import { int, timeNotation } from '../../util/Aliases';
+import { int, timeTyped } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'HOURS\'', () => {
     notation: Notation.Function,
     operation: 'HOURS',
     testTable: `
-    '${timeNotation('11:28:01Z')}' = '${int('11')}'
-    '${timeNotation('15:38:02-08:00')}' = '${int('15')}'
-    '${timeNotation('23:59:00Z')}' = '${int('23')}'
-    '${timeNotation('01:02:03')}' = '${int('1')}'
+    '${timeTyped('11:28:01Z')}' = '${int('11')}'
+    '${timeTyped('15:38:02-08:00')}' = '${int('15')}'
+    '${timeTyped('23:59:00Z')}' = '${int('23')}'
+    '${timeTyped('01:02:03')}' = '${int('1')}'
   `,
   });
 });
