+++ test/integration/functions/op.minutes.test.ts
@@ -1,4 +1,4 @@
-import { int, timeNotation } from '../../util/Aliases';
+import { int, timeTyped } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'MINUTES\'', () => {
     notation: Notation.Function,
     operation: 'MINUTES',
     testTable: `
-    '${timeNotation('11:28:01Z')}' = '${int('28')}'
-    '${timeNotation('15:38:02-08:00')}' = '${int('38')}'
-    '${timeNotation('23:59:00Z')}' = '${int('59')}'
-    '${timeNotation('01:02:03')}' = '${int('2')}'
+    '${timeTyped('11:28:01Z')}' = '${int('28')}'
+    '${timeTyped('15:38:02-08:00')}' = '${int('38')}'
+    '${timeTyped('23:59:00Z')}' = '${int('59')}'
+    '${timeTyped('01:02:03')}' = '${int('2')}'
   `,
   });
 });
