+++ test/integration/functions/op.month.test.ts
@@ -1,4 +1,4 @@
-import { dateNotation, int } from '../../util/Aliases';
+import { dateTyped, int } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'MONTH\'', () => {
     notation: Notation.Function,
     operation: 'MONTH',
     testTable: `
-    '${dateNotation('2010-06-21Z')}' = '${int('6')}'
-    '${dateNotation('2010-12-21-08:00')}' = '${int('12')}'
-    '${dateNotation('2008-06-20Z')}' = '${int('6')}'
-    '${dateNotation('2011-02-01')}' = '${int('2')}'
+    '${dateTyped('2010-06-21Z')}' = '${int('6')}'
+    '${dateTyped('2010-12-21-08:00')}' = '${int('12')}'
+    '${dateTyped('2008-06-20Z')}' = '${int('6')}'
+    '${dateTyped('2011-02-01')}' = '${int('2')}'
   `,
   });
 });
