+++ test/integration/functions/op.seconds.test.ts
@@ -1,4 +1,4 @@
-import { int, timeNotation } from '../../util/Aliases';
+import { int, timeTyped } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'SECONDS\'', () => {
     notation: Notation.Function,
     operation: 'SECONDS',
     testTable: `
-    '${timeNotation('11:28:01Z')}' = '${int('1')}'
-    '${timeNotation('15:38:02-08:00')}' = '${int('2')}'
-    '${timeNotation('23:59:00Z')}' = '${int('0')}'
-    '${timeNotation('01:02:03')}' = '${int('3')}'
+    '${timeTyped('11:28:01Z')}' = '${int('1')}'
+    '${timeTyped('15:38:02-08:00')}' = '${int('2')}'
+    '${timeTyped('23:59:00Z')}' = '${int('0')}'
+    '${timeTyped('01:02:03')}' = '${int('3')}'
   `,
   });
 });
