+++ test/integration/functions/op.timezone.test.ts
@@ -1,4 +1,4 @@
-import { dateNotation, timeNotation } from '../../util/Aliases';
+import { dateTyped, timeTyped } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -16,21 +16,21 @@ describe('evaluation of \'timezone\'', () => {
       "2000-01-01T00:00:00+01:30"^^xsd:dateTime = "PT1H30M"^^xsd:dayTimeDuration
       "2000-01-01T00:00:00-01:30"^^xsd:dateTime = "-PT1H30M"^^xsd:dayTimeDuration
       
-      '${dateNotation('2000-01-01Z')}' = "PT0S"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01+00:00')}' = "PT0S"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01-00:00')}' = "PT0S"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01+00:30')}' = "PT30M"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01+01:00')}' = "PT1H"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01+01:30')}' = "PT1H30M"^^xsd:dayTimeDuration
-      '${dateNotation('2000-01-01-01:30')}' = "-PT1H30M"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01Z')}' = "PT0S"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01+00:00')}' = "PT0S"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01-00:00')}' = "PT0S"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01+00:30')}' = "PT30M"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01+01:00')}' = "PT1H"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01+01:30')}' = "PT1H30M"^^xsd:dayTimeDuration
+      '${dateTyped('2000-01-01-01:30')}' = "-PT1H30M"^^xsd:dayTimeDuration
       
-      '${timeNotation('00:00:00Z')}' = "PT0S"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00+00:00')}' = "PT0S"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00-00:00')}' = "PT0S"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00+00:30')}' = "PT30M"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00+01:00')}' = "PT1H"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00+01:30')}' = "PT1H30M"^^xsd:dayTimeDuration
-      '${timeNotation('00:00:00-01:30')}' = "-PT1H30M"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00Z')}' = "PT0S"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00+00:00')}' = "PT0S"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00-00:00')}' = "PT0S"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00+00:30')}' = "PT30M"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00+01:00')}' = "PT1H"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00+01:30')}' = "PT1H30M"^^xsd:dayTimeDuration
+      '${timeTyped('00:00:00-01:30')}' = "-PT1H30M"^^xsd:dayTimeDuration
       `,
   });
 });
