+++ test/integration/functions/op.year.test.ts
@@ -1,4 +1,4 @@
-import { dateNotation, int } from '../../util/Aliases';
+import { dateTyped, int } from '../../util/Aliases';
 import { Notation } from '../../util/TestTable';
 import { runTestTable } from '../../util/utils';
 
@@ -8,10 +8,10 @@ describe('evaluation of \'YEAR\'', () => {
     notation: Notation.Function,
     operation: 'YEAR',
     testTable: `
-    '${dateNotation('2010-06-21Z')}' = '${int('2010')}'
-    '${dateNotation('2010-12-21-08:00')}' = '${int('2010')}'
-    '${dateNotation('2008-06-20Z')}' = '${int('2008')}'
-    '${dateNotation('2011-02-01')}' = '${int('2011')}'
+    '${dateTyped('2010-06-21Z')}' = '${int('2010')}'
+    '${dateTyped('2010-12-21-08:00')}' = '${int('2010')}'
+    '${dateTyped('2008-06-20Z')}' = '${int('2008')}'
+    '${dateTyped('2011-02-01')}' = '${int('2011')}'
   `,
   });
 });
