+++ test/spec/_data.ts
@@ -1,4 +1,4 @@
-import { dateNotation, dateTimeNotation, decimal, int } from '../util/Aliases';
+import { dateTyped, dateTimeTyped, decimal, int } from '../util/Aliases';
 
 // Data ------------------------------------------------------------------------
 // @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
@@ -43,16 +43,16 @@ export function data() {
     s6: '"abc"^^xsd:string',
     s7: '"DEF"^^xsd:string',
 
-    d1: dateTimeNotation('2010-06-21T11:28:01Z'),
-    d2: dateTimeNotation('2010-12-21T15:38:02-08:00'),
-    d3: dateTimeNotation('2008-06-20T23:59:00Z'),
-    d4: dateTimeNotation('2011-02-01T01:02:03'),
+    d1: dateTimeTyped('2010-06-21T11:28:01Z'),
+    d2: dateTimeTyped('2010-12-21T15:38:02-08:00'),
+    d3: dateTimeTyped('2008-06-20T23:59:00Z'),
+    d4: dateTimeTyped('2011-02-01T01:02:03'),
 
     // TODO: remove this comment: note: this is an illegal data format! date must contain time part
-    dr1: dateNotation('2010-06-21'),
-    dr2: dateNotation('2010-12-21'),
-    dr3: dateNotation('2008-06-20'),
-    dr4: dateNotation('2011-02-01'),
+    dr1: dateTyped('2010-06-21'),
+    dr2: dateTyped('2010-12-21'),
+    dr3: dateTyped('2008-06-20'),
+    dr4: dateTyped('2011-02-01'),
   };
 }
