+++ test/spec/sparql-12/sep-0002/adjust_dateTime-01.spec.ts
@@ -1,5 +1,5 @@
 /* eslint max-len: 0 */
-import { dateTimeNotation, dayTimeDurationNotation } from '../../../util/Aliases';
+import { dateTimeTyped, dayTimeDurationTyped } from '../../../util/Aliases';
 import { Notation } from '../../../util/TestTable';
 import { runTestTable } from '../../../util/utils';
 
@@ -24,11 +24,11 @@ describe('adjust dateTime duration', () => {
       arity: 2,
       notation: Notation.Function,
       testTable: `
-        '${dateTimeNotation('2002-03-07T10:00:00')}' '${dayTimeDurationNotation('-PT10H')}' = '${dateTimeNotation('2002-03-07T10:00:00-10:00')}'
-        '${dateTimeNotation('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationNotation('-PT10H')}' = '${dateTimeNotation('2002-03-07T07:00:00-10:00')}'
-        '${dateTimeNotation('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationNotation('PT10H')}' = '${dateTimeNotation('2002-03-08T03:00:00+10:00')}'
-        '${dateTimeNotation('2002-03-07T00:00:00+01:00')}' '${dayTimeDurationNotation('-PT8H')}' = '${dateTimeNotation('2002-03-06T15:00:00-08:00')}'
-        '${dateTimeNotation('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationNotation('')}' = '${dateTimeNotation('2002-03-07T10:00:00')}'
+        '${dateTimeTyped('2002-03-07T10:00:00')}' '${dayTimeDurationTyped('-PT10H')}' = '${dateTimeTyped('2002-03-07T10:00:00-10:00')}'
+        '${dateTimeTyped('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationTyped('-PT10H')}' = '${dateTimeTyped('2002-03-07T07:00:00-10:00')}'
+        '${dateTimeTyped('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationTyped('PT10H')}' = '${dateTimeTyped('2002-03-08T03:00:00+10:00')}'
+        '${dateTimeTyped('2002-03-07T00:00:00+01:00')}' '${dayTimeDurationTyped('-PT8H')}' = '${dateTimeTyped('2002-03-06T15:00:00-08:00')}'
+        '${dateTimeTyped('2002-03-07T10:00:00-07:00')}' '${dayTimeDurationTyped('')}' = '${dateTimeTyped('2002-03-07T10:00:00')}'
       `,
     });
   });
